requirements.txt
pyautogen
openai